Both sections are independent ADR entries:
- HEAD: 2026-06-10 f2-dual-arm-raw-score-shadowing (#2664)
- main: 2026-06-10 Plan 2.8 Phase-E2 aliasing (#2666)

The Phase-E2 ADR already cross-references #2664 as the sibling bug,
so both belong in the file. Additive, no duplication.

+++ scripts/performance_inference.py
@@ -177,7 +177,7 @@ def sharpe_ci(
     returns: np.ndarray,
     *,
     alpha: float = 0.05,
-    freq: int = 252,
+    freq: float = 252,
     B: int = DEFAULT_B,
     mean_block_length: int = DEFAULT_MEAN_BLOCK_LENGTH,
     method: CIMethod = "studentized",
